put sonar readings in callback function to that other events can be trigerred

+++ ping360_sonar/include/ping360_sonar/sector.h
@@ -52,7 +52,7 @@ public:
     this->radius_sq = half_size*half_size-1;
   }
   void init(float angle, float angle_step);
-  bool nextPoint(int &x, int &y, int &index);
+  bool nextPoint(int &x, int &y, size_t& index);
 };
 
 }
